phpcs: Clean up settings
A rough take on our coding style. Loosening it up a bit here,
tightening it a bit there.

I've commenting out a few rules that should be there at some
point but are too unimportant now.

Let's try to keep it simple first so that we actually pass.
Then we can gradually add enforce new rules in core, add them,
and repeat until we're where we want to be.

Right now there are so many errors that it becomes hard to
reason about anything and actually find regressions / stuff that
wasn't there already.

Change-Id: I371612b39ab3c8663bcf0071d1e09d9a69180ff2