Remove dead code from SpaceBeforeSingleLineCommentSniff
This is dead code that does nothing. Here is why: The regular expression
matches any sequence of hash characters, even an empty sequence. But the
regex is not bound. It matches this anywhere inside the given string.
It searches for a pattern that can be empty. But every string contains an
empty pattern, no matter what the string actually contains. This means this
regex always succeeds and is never false.

I'm aware that this code had a purpose and was introduced in
https://gerrit.wikimedia.org/r/243619 for a reason. However, there was never
a test for this and I'm currently not aware of anybody complaining that this
particular feature was broken. So instead of fixing it I suggest to remove
it for now and re-add it later when somebody really asks for it.

Change-Id: I489874c67e404757f323e66c49f720e8a2d705ee