Speed up PrefixedGlobalFunctionsSniff
This causes 1 scan of all tokens in a file looking
for namespaces where as before one would have been
done for EVERY method in the file.

While profiling on the Title file in mwcore
the tokenIsNamespaced accounted for 22% of execution
time while running phpcs on the file.

After this refactoring tokenIsNamespaced only accounts
for 0.2% of the time in this file.

This is from 2313ms to 23ms.

In its current form of course the static arrays will grow
throughout the phpcs run but I don't really see that becoming
a major issue.
Auto cleanup of the arrays through the phpcs run would be easy
but best not to both until this is needed as we want speed here!

Change-Id: I4cd5cbed47d24a6c3a655e0225d0722f764f7c69