Expect specific output from sniffs
Many of these files were testing multiple things, but the only thing
we asserted was that the file had at least one failure. Patch adds
.expect files for all tests so we can assert that it fails in all the
cases we expect it to fail. I have not manually reviewed the expect
files to determine if all the things that were supposed to be failing
actually do.

.expect files were generated with:

find MediaWiki/Tests/files/ -iname '*.php' -print0 | \
    xargs -0 -I{} bash -c 'vendor/bin/phpcs --standard=MediaWiki {} > {}.expect'

Change-Id: Idd06a613599b2ac9f88bcdb68c558ead29ee2109