Assert fixers do as intended
We didn't have any tests that fixers do what we expect them to do. Patch
adds a test for any .php file that has a file of the same name with
.fixed as a suffix to assert the output of the code matches the recorded
fixed content. If no .fixed file exists we assert that the output of
fixing the file is the source file. I have not manually reviewed these
fixed files, they match the current state of the repository and should
perhaps be manually reviewed.  This was prompted by I3dca4c37 which
tries to fix a bad fixer that was eating code.

This brings in an additional vfs dependency to allow the test to run
without changing anything on the filesystem, but could probably be
rewritten without it if people feel strongly against the dependency.

.fixed files generated with:
vendor/bin/phpcbf --standard=MediaWiki --suffix=.fixed MediaWiki\Tests\files

Change-Id: Ie7102faf842f56e6b93b04f02e90606f19fcaee1