Fix test result parsing, and correct new errors that were exposed
The testFile test compares the expected output of each test with
what it gets from phpcs, and it does this by stripping out irrelevant
output from these two sources before comparing them. The output
in most cases is framed by two lines of hyphens at top and bottom,
and MediaWikiStandardTest::prepareOutput was using these as markers
for the content it was interested in.

However there are various phpcs results that only have one line of
hyphens (such as 'undefined index' errors), and so these were being
discarded.

The fix here is to only discard lines that we know we don't care
about: empty ones, ones with all hyphens, and ones that start with
'Time: ', 'FOUND ', or 'FILE: '. If there are others, they do not
currently show up in any phpcs results (but if they do one day for
new tests, we can deal with them then).

This change to the test highlighted a bunch of existing errors in
the test files, and so these have been fixed. Including the one that
initially prompted this bug report: this involved rearranging when
some fields of tokens are checked, to avoid undefined index errors.

Also remove some unused code and comment (we do check
constructors).

Bug: T146439
Change-Id: I1be4e021e559b2a5efd1c1361a8b574e076e11dc