Update for CodeSniffer 3.0
CodeSniffer 3.0 is a major upstream refactor which requires changes to
all of our custom sniffs and test harness.

Each sniff is now placed in a MediaWiki\Sniffs\<category> namespace, and
uses the namespaced versions of the PHP_CodeSniffer classes.

PrefixedGlobalFunctions is changed to not use static state which is
unreliable in the multi-processed environment.

addWarning() now requires a third $code argument.

The test harness uses the new internal mechanisms for running tests, and
avoids the need for the vfs dependency.

Bug: T142474
Change-Id: I6a934e15fd85aec31b9070c9fbaa34cadd763296