Ignore returns in closures for MissingReturn sniff
If a closure in a function has a `return` statement, but the overall
function doesn't have a `return`, there's no need for a `@return` tag in
the documentation.

The normal findNext() method doesn't have a way of skipping inside a
specific token, so make our own for loop that skips over any closure
inside the function.

Bug: T170765
Change-Id: Id4017ce211bf73727ef9e15e5cc91e08728e06f8