Use local OneClassPerFile sniff for only one class/interface/trait
The separate three upstream sniffs would still have allowed a class and
an interface to be in a file together, which still violates PSR-4.

Temporarily copy in the upstream sniff with the suggested modification
until there is an upstream sniff that also handles this.

See <https://github.com/squizlabs/PHP_CodeSniffer/issues/1627>.

Bug: T173798
Change-Id: Iaae359362aff0b3630f7ca69109b64c64bc50016