Better distinguish "one space before brace" and "brace on same line"
Having the same error message for both checks can result in a warning like:
Expected 1 space before class open brace and should
be same line.find 1

That's confusing, as the "find" part is the same as the first part of the
expected result.

This change creates separate warnings and messages for both checks to better
distinguish the two checks. It also changes the error message from "..same
line.find 1" to a something that looks better: "..same line. Found 1".

Change-Id: I8dda39454389ab01f6c5e84071ba4065674eed06