Remove WhiteSpace.SpaceBeforeSingleLineComment.EmptyComment
Empty comments are useful as paragraph breaks in multi-paragraph comments,
which improves readability of long comments. OTOH there isn't really any
unwanted behavior prevented by this rule that would be likely to occur
in the wild - it is easy get the spacing accidentaly, but accidentally
leaving empty comments is not likely to happen.

I've looked at only exempting empty comments when preceded and followed
by a comment, but PHP_CodeSniffer\Files\File is not really amenable to
line-based operations and it did not seem worth the effort.

Change-Id: I30e99d3e36fe3f18d56bc8fbae1b6e62471d81ea