Be aware of extension MediaWiki compatibility
We can use the MediaWiki requirement specified in extension.json to
customize MediaWiki sniffs for higher accuracy. As a proof of concept, I
adapted the deprecated constant sniff to use it. In the future, we could
have the sniffs that check for PHP 7 syntax run for extensions that are
marked as supporting pre-1.31 versions. This will also be useful for the
DeprecatedGlobal sniff too.

We use composer/semver for the actual parsing of versions to match the
extension.json implementation, with a loose version constraint to not
interfere with MediaWiki core's own dependency (and surely the "semver"
package will actually follow semver ;)).

All of the expensive computations (version parsing) are instance cached
to avoid a major performance impact.

Because the ExtensionInfo class is not a sniff, we can't depend upon
PHPCS's autoloader and need to setup the composer one instead.

Bug: T175465
Change-Id: I726cacda3e92a07483362806f795a25902773496