Fix SpaceyParenthesisSniff comment detection for ignore statements
When changed all the @codingStandardsIgnore in
I92ef235849bcc349c69e53504e664a155dd162c8 I have to move some of them in
tests provider to different line due to this bug. I assumed it in
upstream and therefore not looking in it deeper.
But I found it a problem in our own sniff, so fixing it.

Change-Id: Ia99b628001de7c522db27ca1da706806157fad45