Fix SpaceBeforeControlStructureBraceSniff
The first thing this patch here fixes is re-adding the tests for the
sniff. These have been removed entirely in Ia8b1527, just a few days
after the sniff have been added in Ie23b3e3.

What mainly motivated this patch is the fact that the sniff currently removes
the newline after the final closing brace of an empty "catch {}". This
is because the code contains multiple "+ 2" that jump accross tokens
without checking what is actually skipped. It's assumed to be whitespace,
but if there is a "{}" with no whitespace inbetween, it ends after the
closing brace and starts doing replacements there.

I believe I understand the code I am removing in this patch. I believe the
only reason it is there is to work around an issue further up in the same
code. See, instead of simply replacing the whitespace between the closing ")"
bracket and the opening "{" brace, the code removes the "{", adds it back in
an other position, and forcefully adds a newline, even if there already was
one. The "processEmptyLine" function was added to fix this mistake.

It was not easy to understand all details of the code because no test ever
covered this.

Change-Id: Ic248d52a27e7367f395274ff442a3e5b81966391