Grade errors about duplicate spaces in PHPDoc tags down to warnings
I clearly see the value of a consistent code style. This can even include
disallowing duplicate space characters, like it is done in the cases
touched by this patch. However, I think it is not fair to grade the
following as an "error":

/**
 * @param bool $foo
 * @param int  $bar
 */

Sure, there is a duplicate space. But how does this make this comment any
harder to read?

In contrast, these are actual errors:

/**
 *@param bool $foo
* @param int $bar */

The missing spaces before and after the star. The missing newline at the
end. All this makes the comment hard to read, and all these should and are
reported as errors.

Change-Id: I84e24b5b74a1d3b5e63058ce8bc8ef1b2a9f400f