Remove warning about "missing" @param comments
This particular warning (an error before Ic2aa6c8) makes people (most
notably students trying to make their first commits) add "cargo cult"
cruft like:

 @param string $password Password
 @param ParserOptions $options Parser options
 @param User $user User object

These are literal, real-life examples. Such comments are worse than not
having a comment, because one must read them first to understand they do
not explain anything in addition to what the type and variable name already
explain.

This issue was discussed several times, e.g. in Iee70aad, I2f058e6,
I4257d64, and other places I was not involved in.

I'm very much aware of the positive effect this particular sniff can have
on a codebase where people are interested in having as much documentation
as they can. But I argue the negative effects are not worth it. Especially
because we enforce this rule set on hundreds of not so well maintained
MediaWiki extension codebases on Gerrit. What can people not owning these
codebases do when all they want to do is to "fix" PHPCS "errors"? What can
be said about a string parameter $password other than that it is a password?

Bug: T184650
Change-Id: I125c89e87c332724b046b2384fe0e1aa6f92b9cd