Optimize ClassMatchesFilename sniff for performance
Multiple optimizations are made in this patch. I hope it's ok to have them
in one patch. If you prefer these to be separate patches, please tell me.

Always skip the rest of the file after the first class, interface, or
trait was found. There can only be one match or mismatch against a filename,
because there is only one filename per file. ;-) However, there is one
edge-case when the first class name in a file is a match, but the second is
not. Previously the sniff would fail, but after this change it will succeed.
I find this acceptable because it is not the job of this sniff to work with
multiple classes per file. There is an other sniff for this.

Utilize basename() instead of reimplementing it. I assume this was a mistake.

Turn the loop around that searches for the
"require_once RUN_MAINTENANCE_IF_MAIN;" line in maintenance script.

Change-Id: I28b8849d662ca24ed94c926cf38a7ca99ab2365c