Optimize MultipleEmptyLines sniff for performance
This sniff stops at every whitespace token and repeats the same check,
which involves scanning for the next non-whitespace token, possibly
skipping other whitespace tokens. The fact that these additional
whitespace tokens have already been skipped is forgotten.

We can use this information and tell the engine to not call this sniff
any more on the whitespace tokens that have already been checked.

Change-Id: Id4bb8e7c86bfcc7df7fea222aa4e3240298f7268