Optimize DirUsage sniff for performance
In the profile runs I did this sniff is always among the top 5 most
expensive ones. That's why I believe it's worth looking into the
performance here. In this patch I'm ordering the pre-conditions done
so the most selective ones are executed first.

This sniff is about dirname( __FILE__ ). In all cases where no __FILE__
can be found (which are *all* of them in a clean codebase) the rest of
the conditions is not relevant any more.

So this is how the checks are done now:
* Is the function name "dirname"?
* Is it followed by "("?
* Is it followed by "__FILE__"?
* Is it preceeded by "->", "::", "class ", or "function "? Again, I
  ordered these to have the most selective first.
* Look for the closing ")".

Change-Id: Ie095a5191ad0096ca2b82c7dc8f7505ca6550dd4