Allow @dataProvider annotations in traits
Traits can be used to model certain aspects that should be tested,
similar to a base class that contains a few basic test cases and
providers. Currently this sniffs mistakenly reports this as errors.

I checked the list of "$allowedAnnotations" and believe it's fine to
disallow them all on the "class" level, e.g. when used in the comment
documenting the trait as a whole. That's why I'm intentionally not
touching the second T_CLASS constant in line #212 in this file.

Change-Id: I927705344b49b9afc5c59759b5ffa77a96d2bc8d