Don't report forbidden tags as "should be used inside test classes"
In this patch I'm switching two code paths around:
1. Check if a tag is disallowed. If it is, report this as an error and
shorten out.
2. Check the class name only for allowed tags.

Before this patch the two checks have been done the other way around.
This results in the message "The phpunit annotation … should only be used
inside test classes" for forbidden tags. This sounds like something that
can be fixed. But it can't, because the tag is forbidden, no matter in
which context.

Change-Id: Ica6487a35577f59682351a3998da1f0965d6bfa8