Don't require documenting self-explaining parameter-less functions
Let's say there is a method "public function resetFilePointer()" with no
parameters. What would you write to explain what this method does?
"Resets the file pointer"? :-(

I have seen this happening to often. Similar to the "@param must have a
comment" sniff we recently removed in I125c89e this sniff also makes users
add non-helpful cruft. If a function is badly commented, this must be
pointed out in the original code review where the function was introduced.
It is very hard to come up with useful documentation after the fact.

Since this sniff is heavily used on older codebases where adding PHPDoc
blocks just "for the sake of" adding PHPDoc blocks tends to be
counter-productive, I vote for reducing the impact this sniff does have.

Change-Id: I8b63a5c29ae3bda358cb79af6831cb558e984b9e