Shorten out earlier in the DbrQueryUsage sniff
This patch applies two optimizations:

Instead of scanning for an expected token type and skipping all
unexpected (even non-empty ones), we scan for the previous/next
non-empty token and check if it matches. This does make this sniff
faster and avoids mismatches like in "$dbr->$foo->bar->query".

The two scans are separated now. The second one is not even executed
most of the time. This also speeds this sniff up.

Change-Id: I34c8228fc23545a4f1c42de3c55aa00349aa8a26