Scan for return tags from the end of the function scope
This is based on the fact that a "return" is in many, many cases one of
the very last tokens in a function. Starting the scan from the end should
speed this sniff up.

I made sure inner closures are still skipped the same way as before, both
for performance and because a "return" in a closure must be ignored.

Change-Id: I3f27f8483d0228c5188b8ca6fdb75a76dd7b529f