Optimize ShortCastSyntax sniff for performance
The idea is as follows: The only "content" strings this code can ever
get are the four "(bool)", "(int)", "(boolean)", and "(integer)". The
two short ones are fine, the two longer ones are not.

This patch reduces the average amount of comparisons done from 3 to 1.

I think we will not see an effect in a profiler. But I think it's worth
doing such small optimizations anyway because this code is run so often
on so much code.

Change-Id: Id1aa50f577356f873c197dc42f342cfe192ce775