Optimize PHPUnitClassUsage sniff for performance
If there is no "class" token, the findPrevious() call will scan until
the beginning of the file. This is not a huge issue, because there are
typically not thousands of tokens before an "extends". But it could be a
hundret tokens (typically comments, namespace, and uses).

With this patch the findPrevious() scan also stops when it finds an
"interface" or "trait".

The scan also skips 4 tokens by default. Let's say the current code is
"class Foo extends". The $stackPtr will point at the "extends" token. We
don't need to scan this again (-1). We know there must be a class name
with or without namespace prefixes (-1). We know the class name is
surrounded by spaces (-2). So we know 4 tokens never need to be scanned.

Change-Id: Iee495289dbdc7d72a8562e057142fb343b58f8c8