Make use of $phpcsFile->eolChar in two sniffs
I'm also changing the way the comparison is done. Instead of scanning
via strpos(), the code now does strict comparisons. This works because a
newline always gets it's own token. There is nothing to scan. The newline
tokens are always 1 character long (well, except for Windows, which eolChar
considers).

Because the strings are always 1 character long anyway, this patch can't
have any effect on performance. But I believe it makes the code easier to
read.

Change-Id: Ieb3994c65fe09671cfcf24caf891461f5d2752d9