Fix IllegalSingleLineComment sniff fix for unclosed comments
I found this '\n' in single quotes and tried to understand what it was
meant to represent. It's not a newline but literally the two characters
"\" and "n". I can not think of a reason to have this sequence of characters
somewhere, and replace it with "*/". I'm very sure this was a mistake.

But just replacing the single quotes with double quotes introduced more
problems.

I needed to add two different test cases for this code path that was not
covered before. (Notice the typo in the error code, and how it never appears
in a test.)

I'm not sure if the automatix "fix" that is now done is helpful. Notice how
it potentially enables code that was previously disabled. I will leave this
for the original author(s) to decide. Please feel free to continue working on
this patch!

Change-Id: I79cfbaf72eeb95ddc16eb33a36fc1570bf559065