Optimize PrefixedGlobalFunctions sniff for performance
This is an alternative proposal for Iebd3d8b. The idea is that the
"namespace Foo;" line this loop is looking for can never appear in a scope,
but only on level 0 outside of any other scope. It can not appear inside
classes and functions. It can also not appear inside "namespace Foo { … }"
as all code inside these brackets is already namespaced.

Change-Id: I3b7907d90e5584a0d36363d1a5bda1462e2776ed