Prevent usage of nullable and void type hints
These were introduced in PHP 7.1.

Problematically, PHP 7.0 does not fail lint if a void return type is
used. It also won't fail lint if a nullable parameter type hint is used.
So use PHPCS to prevent those features from being introduced instead. It
will fail lint if a nullable is used as a return type, but that was easy
to implement so it is included for completeness.

Change-Id: I54fa2a3f3904ca62f3a8cb070b650c871cbec8ec