Add possibility to change allowed prefixes
Many extensions use $eg instead of $wg prefixes because of histroical
reasons, thus preventing this sniff to be used. This choice is tough to be
changed as it would break backwards compability in a lot of places and is
considered by the extension.json schema as well with the config_prefix key.

This patch allows to configure an individual list of allowed prefixes to tackle
various use cases.

Bug: T191812
Change-Id: I86949ef4de701093a12ef22932ef2cca89ad18ba