Update message to talk about "top level" instead of "file comment"
Technically, a "file comment" is one that contains the @file PHPDoc tag,
according to the Doxygen documentation and others:
https://www.stack.nl/~dimitri/doxygen/manual/docblocks.html#structuralcommands

I know the wording "file comment" is used different by different people.
It often describes "the first comment block in a file". My suggestion is
to avoid the wording here and talk about what is actually checked: The
nesting level.

Change-Id: Iada159182b1ca6de1345aaef946eab77e66d92e2