Inline token search in UnusedUseStatement sniff
I profiled an example run of the default MediaWiki code sniffer rule set,
and found (again) that this sniff is the by far slowest. About 3 times
slower than the second slowest MediaWiki sniff (that would be the
MultipleEmptyLines one), accounting for 5% of the total runtime of a phpcs
run.

This patch essentially inlines the findNext() call at the bottom of the
while() loop, and replaces all the nested loops with a single for() loop.
The logic is the same as before: only 3 token types are relevant, and each
triggers one or more additional checks in 3 different if() blocks.

The impact is pretty significant: in my example run the 5% mentioned above
got cut down to 2.6%.

This patch does not leave the code in the best readable state, because my
main goal was a most minimal diff. Please feel free to continue working
on this patch if you think the code can be made more readable.

Change-Id: I74d88d64e9255f7d3c788595c4440c99e64047fc