Streamline PHPDoc comment parsing in UnusedUseStatement sniff
This sniff is still the most expensive one in this codebase, accounting for
up to 10% of the total runtime. That's why I had another look and came up
with the proposal in this patch.

The section of code this patch touches is about class names mentioned in
PHPDoc comments like this:

/**
 * @param Class1|Class2 $example Description
 */

The PHPCS parser returns this as follows:
* A T_DOC_COMMENT_TAG token containing the tag "@param".
* A T_DOC_COMMENT_WHITESPACE token containing a single space.
* A T_DOC_COMMENT_STRING token contgaining the entire string
  "Class1|\Namespace\Class2 $example Description".

In this string, we are only interested in the first part before the first
whitespace. We do not care about stuff that is preceeded by a namespace
separator, because thats either part of a namespace or a full qualified
class name that does not need a "use".

The old code did all this using explode(), preg_match_all(), and a string
comparison done in a loop. The new code does the same with a single
preg_match(). The \S makes sure no comparisons are done after the first
whitespace. The (?<!…) is a negative look-behind and makes sure the match
is not preceeded by a namespace separator. The two \b make sure no
substrings are matched. The /i does make the comparison case-insensitive.

The tests prove the code is doing the same as before. I also added a few
more edgy test cases.

Change-Id: Ia7439c25c7bd81b17cdabe6f3f9819889d044f26