Create new sniff for doc comments
Extract the validation code for doc comments from the
FunctionCommentSniff to use the code for all doc comments.
Inside the old sniff it was only running for function comments,
now it also is running for class comments, var comments and file
comments, etc.
It is now also running for function comments on files ending with
Test.php

Bug: T238470
Change-Id: Ic386bc26c29c862cc71c852ca7ebb7c63c1c7d9a