Add rules for PHP71 nullable types
This checks both the signature and the @param annotation to ensure they
both use the same format, and always tries to enforce the PHP71 style.
The only exception is for nullable parameters documented as
`MyClass|null`, which is accepted per T218816.
A warning is also emitted if a typehinted parameter with a default of
'null' is found before a required parameter.

Bug: T231710
Bug: T218816
Change-Id: Id37e6e119711fd5c34d77d36a16f177d5a3e145f