Add a sniff to ensure that setUp and tearDown have :void typehints
Ported from phan plugin at I19b61b91e43d3d4006aa86330b328d1711ae6d7d.
This implementation is a bit more fragile because it checks the class
name and the extended class' name to see if this is a test, as opposed
to parsing the ancestors list. However, the implementation is easier,
and we'll also reach more repos since phan is not used everywhere.

Two issues are added: one is for when the typehint is missing, and is
autofixable. One is for when there's already a typehint !== void. This
is not autofixable because the method has to be refactored manually.

Note: Ideally, this should be the last and only patch to be included in
the next release, so that it will be easier to use PHPCS as a helper for
the PHP7 migration.

Bug: T192167
Change-Id: I0c84dd8a015c012b707ebb4e86f2bec0f47e8bb1