Use severity instead of excludes to allow local overrides
Excluding an error message can not be undone in the .phpcs.xml of a
local codebase. I tried, but couldn't find a way. The severity on the
other hand can be changed back to re-enable the rule, in case a
codebase want's to do this in addition to what the upstream MediaWiki
rule set suggests.

<rule ref="Squiz.WhiteSpace.FunctionSpacing.AfterLast">
	<severity>5</severity>
</rule>
<rule ref="Squiz.WhiteSpace.FunctionSpacing.BeforeFirst">
	<severity>5</severity>
</rule>

Change-Id: Ibf766e34e6ab990dd2218adb97d4e5754a81bd5a