Fix inconsistent sorting in UnsortedUseStatements sniff
The original patch said this would do the same as PHPStorm. I run into
an edge case that doesn't behave the same:

use MediaWiki\Linker\LinkTarget;
use MediaWikiLangTestCase;

This is how PHPStorm sorts it, with the \ always being *before* all
other characters. Unfortunately this is not how natcasesort() behaves
because in the ASCII range the \ is *between* the upper and lower case
characters. This creates confusing results. Example:

Use AAB;
use A\B;
use AaB;

Change-Id: I9a708407c19049efe9e7383b99ad8d447136b23a