phpunit deprecations: Handle assertType() as well
This is almost identical to assertInternalType.

This method should also be deprecated in MediaWiki because there's no
clean replacement for assertInternalType with variable type [1]. Also,
there doesn't seem to be any advantage in using assertType instead of
the appropriate PHPUnit method, regardless of PHPUnit 8.

[1] - Things like $this->assertEquals( $exp, gettype( $actual ) )
*could* work, but again, I don't think we should waste our time
maintaining it.

Also fix a copypasta issue with attribute methods.

Bug: T192167
Change-Id: I8ef556b630812aeea77c5606713f53d9af609f1b