Upgrade to PHPUnit 8
This updates a few attributes in phpunit.xml.dist that were no longer
valid.

Note that enforceTimeLimit (replacing beStrictAboutTodoAnnotatedTests)
will not work without the phpunit/php-invoker and prints errors.
It does still run the tests though. See T192167#5722991.

I did not add phpunit/php-invoker as it depends on ext-pcntl, as it
does not work on windows, which would prevent me from creating this
change.

Change-Id: Ie9c1b8c2c160c30dd566e283b327213e1815b83d