Fix sniff destroying comments mentioning other variables
This is the most minimal fix I can think of to make this sniff stop destroying some
otherwise valid @param tags. The only relevant change is the addition of the
previously missing ^ character to force the regular expression to only match at the
beginning of the string, not somewhere else in the middle of a longer comment.

Additional changes:
* I replaced (?:a|b) with [ab], which is the same for single characters.
* I replaced [^\s] with \S, which is the same.
* I swapped [&$] to reflect how the characters are actually used in code.

Bug: T174761
Change-Id: If0d37dadc0d5cec43e38ac890c443716b1fc8ba8