Rewrite MultipleEmptyLines sniff for performance
This is one of the slowest sniffs we have, for relatively little value.
The reason for the bad performance is simple: There is just so many
whitespace in our code, and this sniff gets triggered for every whitespace
token. There was not really something wrong with the code. Still we can do
better.

This rewrite cuts the runtime for this sniff down to about a third,
according to my XDebug profiling. Main reason for this is because this
rewrite skips the most common sequence of whitespace tokens (newline + tabs
at the begining of each line), and does not trigger the sniff again for the
tabs.

This rewrite comes with a notable difference: Now the sniff assumes there
is no horizontal whitespace at the end of a line. There is another sniff
that checks and enforces this already. I added a test case for this
situation.

Change-Id: Iee09735de2f57f38e0b0940d033d1c1737f4909a