Add problematic values 1 and 1.0 to PHPUnitAssertEqualsSniff
PHPUnit's assertEquals() considers all these values to be equal:
true
1
1.0
'1'
'1.0'
'1.00'
And so on.

What internally happens is: Cast both values to string; check if both
are numeric; cast both to numbers; compare these numbers.

This is confusing and almost always not what was intended. Enforce to
use the strict assertSame() instead. In case an assertion was meant to
do a relaxed float comparison, it should use assertEqualsWithDelta().

Bug: T246662
Change-Id: I735ab8b54e71504bfa4d89bd9ae8b61748fd9eeb