Report particularly confusing assertNotEquals( false )
This adds a small set of very specific assertions to the list:

* assertNotEquals() as well as assertNotSame() with a boolean value is
not wrong, but highly confusing. Replace with assertTrue/False().

* Please note that it would be a bad idea to replace other
assertNotEquals() with assertNotSame(). assertNotSame() is not better.
It does have the exact same issues as assertNotEquals(). A naive
replacement would make the assertion *less* strict, the opposite of what
this sniff aims for. Example:
  * assertNotEquals( true ) reports pretty much *everything*, except
    values that are "falsy", e.g. null, 0, or the empty string.
  * assertNotSame( true ) reports *only* true. All other values that
    would have been reported by assertNotEquals() are accepted.

Same problem with assertNotNull() and such. Let's please never encourage
people to use these.

Change-Id: I9e91697dfd71f93eb465c2c4af3cc66abec2a8fe