Add sniff for deprecated $wgMemc
It can be replaced with services, which means that the replacement is easy enough to add this sniff (like for $wgContLang).

Also, this isn't much used in WMF-deployed code: https://codesearch.wmflabs.org/deployed/?q=%5C%24wgMemc%5Cb&i=nope&files=&repos=

The global is deprecated since 1.35, so I don't object to waiting for
said version to be released before merging this patch.

Bug: T160813
Change-Id: I496e37c33558a63c546439405e634a56c30a7bfb