New error: Expected parameter type before parameter name
Relax the regex to match empty string before the parameter name and
check for the empty string
Skip the type_space check: no type, no spaces between type and name

This also trigger further checks for the parameter name like the
punctation check in the tests. Before the whole line was skipped and
other checks detects the param as missing.

Bug: T141412
Change-Id: I16c64e614bc57314157cc641568921b8c639b6b3