Fix minor performance issue in UnsortedUseStatementsSniff
Before, the stack pointer returned at the end pointed *at* the last
use statement in the list. This single use statement triggered the sniff
a second time. This additional call had no effect because there is no
way to reorder a list that contains a single element only. Now, this
pointless call is skipped, reducing the total runtime of this sniff.

This patch also fixes and improves some PHPDoc documentation in the
sniff.

Change-Id: I340cf44baf2d3a3c7a37e9642e44a3d4d1bab89d