Full rewrite of the UnusedUseStatementSniff, 4x faster
This sniff is the by far slowest in this codebase. (The only one that's
slower is the SuperfluousWhitespaceSniff from the Squiz ruleset.)

This is because the sniff is iterating the entire file for every
individual use statement. The more use statements are in a file, the
more iterations are done. This adds up quickly, no matter how fast the
individual operations are.

This patch implements a different approach: First, it scans for a
*group* of use statements. Then, it does the same loop over the entire
file, but looks for more than one class name. If a usage is found, it's
removed from the list and never checked for again. Which means the loop
gets faster and faster towards the end.

Multiple groups of use statements are possible separated by non-empty
tokens. That's rare. The sniff will run multiple times then, one time for
each group.

In my example test run on a medium size codebase the difference is as
follows (before vs. after):

Runtime of this sniff alone, according to XDebug:
544.000 vs. 148.000 ticks (down to 27%)

How often this sniff is called:
870 vs. 160 times
That's expected because of the different approach.

Relative runtime of this sniff when run together with all other sniffs:
2.4% vs. 0.7%
That's much better than I was hoping for.

Change-Id: Ia4e97983db247f2eb2748d8c87dedf717b7c171d