Significant simplification of UnusedGlobalVariables sniff
Instead of remembering all variable names, we can just remove a known
global from the list. What's left in the list is what's unused.

This assumes that a global can not be used before it is declared. This
is indeed the case in PHP.

unset() works fine when the element does not exist. Just doing a
potentially pointless unset() is faster than checking for the elements
presence before.

Change-Id: Ib477bfa300c7761c685a82ff8f83d1e1b9e8efe0