UnusedUseStatementSniff: Add @property-read and @property-write
These are standard annotations (phpDocumentor, PSR19 draft) used to
indicate that a property can only be respectively read from or written
to.

They're uncommon because @property is usually enough, but they might
make a difference in some cases (e.g. with phan's dead code detection),
and they're used in a handful or repos according to codesearch.

Also add them to a test, and add @property to the test as well.

Change-Id: Ia06ebd9f94a7a7d6d5f55d4840c892a285e71152