Fix UnsortedUseStatements sniff failing with leading backslashes
Leading backslashes are removed by the PSR12.Files.ImportStatement
sniff. Unfortunately the current implementation of the
UnsortedUseStatements sniff gets confused and reports this as not
being sorted, even if it is. This creates a confusing, unnecessary
error message.

This patch also adds test cases with leading backslashes to the
UnusedUseStatement sniff. The sniff is already fine with this.

Change-Id: I05f7fab50accde62b32902fd10881988c195d79a