Recognize used classes even if variable/type in a tag are flipped
This can happen in both:
* @param $param Class
* /** @var $var Class */

Variable name and type are flipped in these examples. This violates our
style guide. However, it's not the job of *this* sniff to report this
formatting error.

Before, these examples resulted in classes being reported as "unused",
which is highly confusing.

This is a resubmission of I355ddeb with the same test case, but a
different fix.

Bug: T253711
Change-Id: I60bb40327f0059edba304623492b24f71ce2f69d