Add RedundantVarNameSniff from WikibaseCodeSniffer
Note there are 2 types of @var comments. If it's inline, the variable
name must be repeated:

/** @var int $x */
$x = foo();

But this is pointless for class properties:

/**
 * @var int $x
 */
private $x;

Copied from
https://github.com/wmde/WikibaseCodeSniffer/blob/master/Wikibase/Sniffs/Commenting/RedundantVarNameSniff.php
and slightly modified, e.g. to be compatible with typed properties in
PHP 7.4.

Just to be sure I did a quick benchmark and found the runtime of this
additional sniff is neglegible.

Bug: T164653
Change-Id: Ic2aae66a21a7272d9a264cdb7ed6238755a8eda6