Add (optional) FullQualifiedClassNameSniff from WikibaseCodeSniffer
A custom sniff that enforces to use "use" statements, instead of using
full qualified class names in the code.

This is a rewrite of a custom sniff originally developed for the
Wikibase CodeSniffer. Since this is a code style we can't and shouldn't
enforce on all codebases (in my opinion), it's disabled by default.

Bug: T164653
Change-Id: I6cb64a7878a2b00017ad75ee479d159baf19781d