Fix ValidGlobalNameSniff possibly running in an endless loop
This can happen when PHPCS either runs in the background while coding,
or when PHPCS is manually executed, but a file is still being worked on
and not completed yet.

This particular sniff runs into an actual endless loop. This is because
the ++ operator does not do anything if the variable contains false. It
stays false forever.

I believe we should have a test case for this. To be able to have a test
we can not expect all test files to be valid any more. Sure, usually they
are (because we usually run parallel-lint before we run phpcs). But this
is not guaranteed. Especially not when using PHPCS in a local dev
environment.

Change-Id: I23b6aa58fe1b91fd45d816269f3462660388887e