Fix possible index error in ReferenceThisSniff
This can happen when PHPCS either runs in the background while coding,
or when PHPCS is manually executed, but a file is still being worked on
and not completed yet.

I believe we should have a test case for this. To be able to have a test
we can not expect all test files to be valid any more. Sure, usually they
are (because we usually run parallel-lint before we run phpcs). But this
is not guaranteed. Especially not when using PHPCS in a local dev
environment.

Change-Id: I4d59aaa526d722b126c4365e362825cfb9ffec9d