Fix handling of comment string for MissingReturnType
After addition of the error the checks are processed.
When the token is the close token this results in a "Return type should
not end with punctuation", on autofix the / from the close token is
removed and for the next sniff this is a longer comment adding * on each
line.
Rewrite the sniff to just use the content when the correct token.

Change-Id: I1468814a9f6081af36f74354996c6182fa0525bd